+++ src/ck-core/ckarray.h
@@ -431,7 +431,7 @@ extern void (*ArrayMsgRecvExtCallback)(int, int, int *, int, int, char *, int);
 extern int (*ArrayElemLeaveExt)(int, int, int *, char**, int);
 extern void (*ArrayElemJoinExt)(int, int, int *, int, char*, int);
 extern void (*ArrayResumeFromSyncExtCallback)(int, int, int *);
-extern void (*ArrayMsgGPUDirectRecvExtCallback)(int, int, int*, int, int, long*, void *, int, char*, int);
+extern void (*ArrayMsgGPUDirectRecvExtCallback)(int, int, int*, int, int, int*, void *, int, char*, int);
 
 class ArrayElemExt: public ArrayElement {
 private:
@@ -456,7 +456,7 @@ public:
     if (CMI_ZC_MSGTYPE((char *)UsrToEnv(impl_msg)) == CMK_ZC_DEVICE_MSG) {
         int numDevBufs; implP | numDevBufs;
         int directCopySize; implP | directCopySize;
-        long devBufSizes[numDevBufs];
+        int devBufSizes[numDevBufs];
 
         CkDeviceBuffer *devBufs = new CkDeviceBuffer[numDevBufs];
         for (int i = 0; i < numDevBufs; i++) {
